# metlink-train-schedule-nextjs/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Metlink Train Schedule API
  description: API for accessing real-time train departure information for Wellington train lines
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: https://your-domain.com/api/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Development server

paths:
  /departures:
    get:
      summary: Get train departures
      description: Retrieve real-time train departure information for specified stations and line
      operationId: getDepartures
      tags:
        - Departures
      parameters:
        - name: stations
          in: query
          description: Comma-separated list of station codes (e.g., WELL,PETO,FEAT)
          required: false
          schema:
            type: string
            example: "WELL,PETO,FEAT"
        - name: line
          in: query
          description: Train line code (WRL, KPL, HVL, JVL)
          required: false
          schema:
            type: string
            enum: [WRL, KPL, HVL, JVL]
            default: WRL
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeparturesResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /analytics/performance:
    get:
      summary: Get performance statistics
      description: Retrieve API performance metrics
      operationId: getPerformanceStats
      tags:
        - Analytics
      parameters:
        - name: endpoint
          in: query
          description: Filter by endpoint
          required: false
          schema:
            type: string
        - name: startDate
          in: query
          description: Start date for metrics (ISO 8601)
          required: false
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: End date for metrics (ISO 8601)
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformanceStatsResponse'

  /analytics/on-time-performance:
    get:
      summary: Get on-time performance metrics
      description: Retrieve on-time performance statistics for a service line
      operationId: getOnTimePerformance
      tags:
        - Analytics
      parameters:
        - name: serviceId
          in: query
          description: Service ID (WRL, KPL, HVL, JVL)
          required: true
          schema:
            type: string
            enum: [WRL, KPL, HVL, JVL]
        - name: startDate
          in: query
          description: Start date (ISO 8601)
          required: false
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: End date (ISO 8601)
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnTimePerformanceResponse'

  /analytics/historical:
    get:
      summary: Get historical departure data
      description: Retrieve historical departure records for analytics
      operationId: getHistoricalData
      tags:
        - Analytics
      parameters:
        - name: serviceId
          in: query
          description: Service ID filter
          required: false
          schema:
            type: string
        - name: stopId
          in: query
          description: Stop ID filter
          required: false
          schema:
            type: string
        - name: station
          in: query
          description: Station code filter
          required: false
          schema:
            type: string
        - name: startDate
          in: query
          description: Start date (ISO 8601)
          required: false
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: End date (ISO 8601)
          required: false
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Maximum number of records to return
          required: false
          schema:
            type: integer
            default: 1000
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoricalDataResponse'

components:
  schemas:
    Departure:
      type: object
      properties:
        service_id:
          type: string
          description: Service line identifier
        destination:
          type: object
          properties:
            stop_id:
              type: string
            name:
              type: string
        departure:
          type: object
          properties:
            aimed:
              type: string
              format: date-time
            expected:
              type: string
              format: date-time
        status:
          type: string
        stop_id:
          type: string
        station:
          type: string

    DeparturesResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            inbound:
              type: array
              items:
                $ref: '#/components/schemas/Departure'
            outbound:
              type: array
              items:
                $ref: '#/components/schemas/Departure'
            total:
              type: integer
            version:
              type: string
        meta:
          type: object
          properties:
            cached:
              type: boolean
            cacheAge:
              type: string

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string

    PerformanceStatsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            total:
              type: integer
            averageResponseTime:
              type: integer
            p50:
              type: integer
            p95:
              type: integer
            p99:
              type: integer
            errorRate:
              type: number
            statusCodes:
              type: object
              additionalProperties:
                type: integer

    OnTimePerformanceResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            total:
              type: integer
            onTime:
              type: integer
            delayed:
              type: integer
            cancelled:
              type: integer
            averageDelay:
              type: number
            onTimePercentage:
              type: number

    HistoricalDataResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            departures:
              type: array
              items:
                type: object
            count:
              type: integer